requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torch-geometric>=2.3.0

# Optional graph extensions (install only if torch-geometric fails)
# pip install torch-scatter torch-sparse --index-url https://data.pyg.org/whl/
# torch-scatter>=2.1.0
# torch-sparse>=0.6.17

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
PyWavelets>=1.4.1

# Image processing
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Data
pandas>=2.0.0

# Utils
umap-learn>=0.5.3
tqdm>=4.65.0
joblib>=1.3.0
requests>=2.31.0

# Data download (optional)
kaggle>=1.5.13
